rewrite gemmini conv
